+++ test/domain.test.js
@@ -7,26 +7,23 @@ const Internal = protocol.getDomain('Internal');
 
 describe('Domain', () => {
 
-    describe('createInputsFromObject', () => {
+    describe('prepareInputs', () => {
 
         it('parses and validate object', async () => {
-            const inputs = await Internal.createInputsFromObject({
+            const inputObject = {
                 url: 'http://example.com',
                 foo: 42,
-                options: {},
+            };
+            const inputs = await Internal.prepareInputs({
+                inputObject,
+                applyDefaults: false,
             });
-            expect(inputs.length).toEqual(3);
+            expect(inputs.length).toEqual(2);
             const url = inputs.find(i => i.key === 'url');
             expect(url).toExist();
             expect(url.key).toEqual('url');
             expect(url.data).toEqual('http://example.com');
             expect(url.valid).toEqual(true);
-            expect(url.errors.length).toEqual(0);
-            const options = inputs.find(i => i.key === 'options');
-            expect(options).toExist();
-            expect(options.key).toEqual('options');
-            expect(options.data).toEqual({ flag: false });
-            expect(options.valid).toEqual(true);
             expect(url.errors.length).toEqual(0);
             const foo = inputs.find(i => i.key === 'foo');
             expect(foo).toExist();
@@ -37,8 +34,12 @@ describe('Domain', () => {
         });
 
         it('applies default inputs', async () => {
-            const inputs = await Internal.createInputsFromObject({
+            const inputObject = {
                 url: 'http://example.com',
+            };
+            const inputs = await Internal.prepareInputs({
+                inputObject,
+                applyDefaults: true,
             });
             expect(inputs.length).toEqual(2);
             const options = inputs.find(i => i.key === 'options');